gah, what a mess - the last two updates to this package were broken, since el5 provides libxslt 1.1.17 and XML::LibXSLT > 1.63 requires libxslt >= 1.1.18.  rolling back to 1.63 since this package will never build in its current state.
oh, and just for more fun, XML::LibXML 1.70 introduces a non-backwards-compatible ABI, so this package can't work with XML::LibXML >= 1.70.  to get this package to build we therefore need perl-XML-LibXML-1.69-2 to build, which also means pulling perl-XML-LibXML-Common out of the vault (since it was absorbed into XML::LibXML in, yes, 1.70).

oy vey.  thank you, David Steinbrunner, for bringing this issue to our attention :)