initial import of upstream icinga-web package 1.7.0
this has been extensively tested by community members and developers,
requiring icinga-web to become even more packagable.

as icinga itsself, the database setup must be done manually as well.

there is no implicit dependency on the icinga package, as icinga-web
will use the idoutils database, once configured in databases.xml - but
that can be on a different host.

the configs will be placed into /etc/icinga-web/conf.d/*.xml and not
overriden on upgrades. After install/upgrade a config cache clearing is
invoked, as internal agavi strctures require that. the cache itsself
lives in /var/cache/icinga-web while the logs run into
/var/log/icinga-web - these are web logs and debug logs, apache log is
different.

if icinga-web is installed on the same host as icinga with idoutils the
package sets various defaults, such as the external command file within
/var/spool/icinga/cmd/icinga.cmd (newly introduced in icinga 1.7.0
rpms). furthermore, the icinga binary and config file is referenced as
the icinga package will do.

the apache config will be installed into
/etc/httpd/conf.d/icinga-web.conf where the (rewrite) magic happens.

currently the package does not require php53 on el5, but it could happen
in the future, as the requirement is php 5.2.3 the least currently (and
5.3 is even better).

if you have a pnp4nagios install, and want to integrate that into
icinga-web the module-pnp package will help. This will install the
required cronk xml snippets in order to add pnp urls. you will find the
final install in /usr/share/icinga-web (where icinga-web app lives)
underneath app/modules/Cronks/data/xml/extensions/

since this is the initial push to an official repo, please report any
bugs from the package directly to https://dev.icinga.org as the spec
file is kindly the same as provided with the upstream tarball.